+++ packages/image_picker_plus_ios/pigeons/messages.dart
@@ -7,13 +7,7 @@ import 'package:pigeon/pigeon.dart';
 @ConfigurePigeon(
   PigeonOptions(
     dartOut: 'lib/src/messages.g.dart',
-    objcHeaderOut:
-        'ios/image_picker_ios/Sources/image_picker_ios/include/image_picker_ios/messages.g.h',
-    objcSourceOut: 'ios/image_picker_ios/Sources/image_picker_ios/messages.g.m',
-    objcOptions: ObjcOptions(
-      prefix: 'FLT',
-      headerIncludePath: './include/image_picker_ios/messages.g.h',
-    ),
+    swiftOut: 'ios/image_picker_ios/Sources/Messages.swift',
     copyrightHeader: 'pigeons/copyright.txt',
   ),
 )
@@ -60,7 +54,7 @@ class SourceSpecification {
 @HostApi()
 abstract class ImagePickerApi {
   @async
-  @ObjCSelector('pickImageWithSource:maxSize:quality:fullMetadata:')
+  @SwiftFunction('pickImage(withSource:maxSize:quality:fullMetadata:)')
   PickedMedia? pickImage(
     SourceSpecification source,
     MaxSize maxSize,
@@ -68,7 +62,7 @@ abstract class ImagePickerApi {
     bool requestFullMetadata,
   );
   @async
-  @ObjCSelector('pickMultiImageWithMaxSize:quality:fullMetadata:limit:')
+  @SwiftFunction('pickMultiImage(withMaxSize:quality:fullMetadata:limit:)')
   List<PickedMedia> pickMultiImage(
     MaxSize maxSize,
     int? imageQuality,
@@ -76,14 +70,14 @@ abstract class ImagePickerApi {
     int? limit,
   );
   @async
-  @ObjCSelector('pickVideoWithSource:maxDuration:')
+  @SwiftFunction('pickVideo(withSource:maxDuration:)')
   PickedMedia? pickVideo(SourceSpecification source, int? maxDurationSeconds);
   @async
-  @ObjCSelector('pickMultiVideoWithMaxDuration:limit:')
+  @SwiftFunction('pickMultiVideo(withMaxDuration:limit:)')
   List<PickedMedia> pickMultiVideo(int? maxDurationSeconds, int? limit);
 
   /// Selects images and videos and returns their paths.
   @async
-  @ObjCSelector('pickMediaWithMediaSelectionOptions:')
+  @SwiftFunction('pickMedia(withMediaSelectionOptions:)')
   List<PickedMedia> pickMedia(MediaSelectionOptions mediaSelectionOptions);
 }